feat: basic translation search
It's now possible to search for a specific text or text fragment. The kook up is performed either in translated strings or its corresponding key.

A "nothing found" message is shown if the search string doesn't match anything.

A search for an empty string "" returns a paginated list of all translations.

Closes #1